refactor: break up bundle analysis services
These changes add nothing. It's a simple refactor breaking up
different services related to bundle analysis into different files.

The "bundle_analysis" service is already a grouping of different things
(processing, comparison, notify, etc) and it's getting big as is.
